[SSR] Insert dehydrated data directly into HTML doc after main markup.
This reverts commit 42cbf810a0c7f294042110a91a54d0af5668fba6, which
reverted https://github.com/artsy/force/pull/3108
