+++ src/desktop/apps/collect2/server.tsx
@@ -1,17 +1,18 @@
-import React from "react"
 import { buildServerApp } from "reaction/Artsy/Router/server"
 import { stitch } from "@artsy/stitch"
-import { routes } from "reaction/Apps/Collections/routes"
+import { routes } from "reaction/Apps/Collect/routes"
 import express from "express"
+import React from "react"
+import { Meta } from "./meta"
 import { buildServerAppContext } from "desktop/lib/buildServerAppContext"
 
-const app = (module.exports = express())
+export const app = express()
 
-app.get("/collections", async (req, res, next) => {
+const index = async (req, res, next) => {
   try {
-    const { IS_MOBILE } = res.locals.sd
+    const { APP_URL, IS_MOBILE } = res.locals.sd
 
-    const { ServerApp, headTags, redirect } = await buildServerApp({
+    const { headTags, ServerApp, redirect, scripts } = await buildServerApp({
       routes,
       url: req.url,
       context: buildServerAppContext(req, res),
@@ -30,13 +31,14 @@ app.get("/collections", async (req, res, next) => {
         styledComponents: true,
       },
       blocks: {
-        head: () => <React.Fragment>{headTags}</React.Fragment>,
-        body: ServerApp,
+        head: () => <Meta appUrl={APP_URL} headTags={headTags} />,
+        body: () => <ServerApp />,
       },
       locals: {
         ...res.locals,
-        assetPackage: "collections",
+        assetPackage: "collect2",
         bodyClass: IS_MOBILE ? "body-header-fixed body-no-margins" : null,
+        scripts,
       },
     })
 
@@ -45,6 +47,8 @@ app.get("/collections", async (req, res, next) => {
     console.log(error)
     next(error)
   }
-})
+}
 
-export default app
+app.get("/collect", index)
+app.get("/collect/:medium?", index)
+app.get("/collection/:slug", index)
