+++ src/desktop/apps/collections/server.tsx
@@ -1,19 +1,17 @@
+import React from "react"
 import { buildServerApp } from "reaction/Artsy/Router/server"
 import { stitch } from "@artsy/stitch"
-import { routes } from "reaction/Apps/Collect/routes"
+import { routes } from "reaction/Apps/Collections/routes"
 import express from "express"
-import React from "react"
-import { Meta } from "./meta"
 import { buildServerAppContext } from "desktop/lib/buildServerAppContext"
 
-const app = (module.exports = express())
+export const app = express()
 
-const index = async (req, res, next) => {
+app.get("/collections", async (req, res, next) => {
   try {
-    const user = req.user && req.user.toJSON()
-    const { APP_URL, IS_MOBILE } = res.locals.sd
+    const { IS_MOBILE } = res.locals.sd
 
-    const { headTags, ServerApp, redirect } = await buildServerApp({
+    const { ServerApp, headTags, redirect, scripts } = await buildServerApp({
       routes,
       url: req.url,
       context: buildServerAppContext(req, res),
@@ -32,13 +30,14 @@ const index = async (req, res, next) => {
         styledComponents: true,
       },
       blocks: {
-        head: () => <Meta appUrl={APP_URL} headTags={headTags} />,
-        body: () => <ServerApp />,
+        head: () => <>{headTags}</>,
+        body: ServerApp,
       },
       locals: {
         ...res.locals,
-        assetPackage: "collect2",
+        assetPackage: "collections",
         bodyClass: IS_MOBILE ? "body-header-fixed body-no-margins" : null,
+        scripts,
       },
     })
 
@@ -47,10 +46,4 @@ const index = async (req, res, next) => {
     console.log(error)
     next(error)
   }
-}
-
-app.get("/collect", index)
-app.get("/collect/:medium?", index)
-app.get("/collection/:slug", index)
-
-export default app
+})
