+++ src/desktop/components/main_layout/templates/react_redesign.jade
@@ -9,12 +9,15 @@ block head
   != css
 
 block body
-  script.
-    var __BOOTSTRAP__ = !{JSON.stringify(data)}
-
   #react-root
     != body
   #react-portal
 
   if jsonLD
     include ./json_ld
+
+//- All hydration data should be loaded before React et al
+block prepend scripts
+  script.
+    var __BOOTSTRAP__ = !{JSON.stringify(data)}
+  != scripts
