+++ src/desktop/components/main_layout/templates/redesign.jade
@@ -38,10 +38,12 @@ html(
 
     //- Javascripts
     #scripts
-      include scripts
-      script( src= asset('/assets/main_layout.js') )
-
-      if assetPackage
-        script( src=asset('/assets/#{assetPackage}.js') )
-
+      //- Don't replace this block, rather prepend or append!
       block scripts
+
+        include scripts
+        script( src= asset('/assets/main_layout.js') )
+
+        if assetPackage
+          script( src=asset('/assets/#{assetPackage}.js') )
+
