+++ src/desktop/index.js
@@ -30,8 +30,8 @@ app.use(require("./apps/artwork"))
 app.use(require("./apps/artwork2/server").app)
 app.use(require("./apps/about"))
 app.use(require("./apps/collect_art"))
-app.use(require("./apps/collect2/server"))
-app.use(require("./apps/collections/server"))
+app.use(require("./apps/collect2/server").app)
+app.use(require("./apps/collections/server").app)
 app.use(require("./apps/categories"))
 app.use(require("./apps/consign"))
 app.use(require("./apps/contact"))
@@ -94,4 +94,4 @@ app.use(require("./apps/user"))
 
 // Examples
 app.use(require("./apps/react_example"))
-app.use(require("./apps/isomorphic-relay-example/server"))
+app.use(require("./apps/isomorphic-relay-example/server").app)
